-Fixed a CalibrateMonSpd bug and added a new option to bitsEncodeDIO to prevent it from loading the gamma table every time.
git-svn-id: https://psychtoolbox-3.googlecode.com/svn/trunk@714 f5efbe48-b203-0410-9429-fc5e098e5a54

+++ Psychtoolbox/PsychCal/CalibrateMonDrvr.m
@@ -152,7 +152,7 @@ for a = 1:cal.describe.nAverage
     for i = 1:cal.nDevices
         disp(sprintf('Monitor device %g',i));
         Screen('FillRect', window, 1, boxRect);
-        Screen('Flip', window, 0, g_usebitspp);
+        Screen('Flip', window, 0, double(g_usebitspp));
 
         % Measure ambient
         darkAmbient1 = MeasMonSpd(window, [0 0 0]', cal.describe.S, 0, whichMeterType, theClut);
